# Java/swagger.yml
swagger: "2.0"
info:
  title: Swagger Hire-it
  version: '1.0'
...
paths:
...
  /demandas:
    post:
      tags:
      - Demanda
      summary: "Cria uma demanda"
      description: "Cria uma demanda com tecnologias e softskills"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/Demanda"
      responses:
        '201':
          description: Created
        '422':
          description: Informações inválidas
    get:
      tags:
      - Demanda
      summary: "Retorna todas demandas"
      description: "Retorna todas demandas"
      responses:
        '200':
          description: Ok
          schema:
            $ref: "#/definitions/DemandaResponse"
        '204':
          description: No content
  /demandas/{idDemanda}:
    get:
        tags:
        - Demanda
        summary: "Retorna uma demanda"
        description: "Retorna uma demanda de acordo com o ID no Path"
        parameters:
        - name: "idDemanda"
          in: "path"
          description: "ID da demanda que deve retornar"
          required: true
          type: "integer"
          format: "int64"
        responses:
          '200':
            description: Ok
            schema:
              $ref: "#/definitions/DemandaIdResponse"
          '204':
            description: No content
  /demandas/usuario/{idUsuario}:
    get:
        tags:
        - Demanda
        summary: "Retorna todas demandas de um usuário"
        description: "Retorna todas demandas de um usuário com o ID no Path"
        parameters:
        - name: "idUsuario"
          in: "path"
          description: "ID do usuario"
          required: true
          type: "integer"
          format: "int64"
        responses:
          '200':
            description: Ok
            schema:
              $ref: "#/definitions/DemandaUsuarioResponse"
          '204':
            description: No content
  /demandas/filtro:
    get:
        tags:
        - Demanda
        summary: "Retorna todas demandas filtradas"
        description: "Retorna todas demandas com base nas informações passadas no filtro"
        parameters:
          - in: "body"
            name: "body"
            description: "Todos os campos são opcionais"
            required: false
            schema:
              $ref: "#/definitions/DemandaFiltro"
        responses:
          '200':
            description: Ok
            schema:
              $ref: "#/definitions/DemandaFiltroResponse"
          '204':
            description: No content
...
  /export/demanda:
    post:
      tags:
      - Exportacao
      summary: "Exporta demandas"
      description: "Faz uma exportação das demandas do usuário para uma lista"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/ExportacaoDemanda"
      responses:
        '200':
          description: Ok 
        '400':
          description: Erro ao criar a demanda    
...
  /ofertas/{idOferta}:
    get:
      tags:
      - Oferta
      summary: "Retorna uma oferta"
      description: "Retorna uma oferta com base no ID"
      parameters:
        - name: "idOferta"
          in: "path"
          description: "ID da oferta que deve retornar"
          required: true
          type: "integer"
          format: "int64"
      responses:
        '200':
          description: Ok
          schema:
            $ref: "#/definitions/OfertaIdResponse"
        '204':
          description: No content
...
  /usuario/{idUsuario}:
    get:
        tags:
        - Usuário
        summary: "Retorna um usuário"
        description: "Retorna um usuário de acordo com o ID no Path"
        parameters:
        - name: "idUsuário"
          in: "path"
          description: "ID do usuário que deve retornar"
          required: true
          type: "integer"
          format: "int64"
        responses:
          '200':
            description: Ok
            schema:
              $ref: "#/definitions/UsuarioIdResponse"
          '204':
            description: No content
    delete:
      tags:
        - Usuário
      summary: "Deleta um usuário com base em um ID"
      description: "Deleta um usuário com base em um ID"
      parameters:
        - name: "idUsuario"
          in: "path"
          description: "ID do usuário que deve ser deletado"
          required: true
          type: "integer"
          format: "int64"
      responses:
        '204':
          description: No content
        '404':
          description: Not Found
    put:
      tags:
        - Usuário
      summary: "Atualiza um usuário com base em um ID"
      description: "Atualiza um usuário com base em um ID"
      parameters:
        - name: "idUsuario"
          in: "path"
          description: "ID do usuário que deve ser atualizado"
          required: true
          type: "integer"
          format: "int64"
        - in: "body"
          name: "body"
          schema:
            $ref: "#/definitions/Usuario"
      responses:
        '204':
          description: No content
        '404':
          description: Not Found
        '422':
          description: Informações inválidas
...
definitions:
...
    Demanda:
      type: "object"
      properties:
        demandas:
          properties:
            titulo:
              type: "string"
              example: "Primeira demanda"
            descricao:
              type: "string"
              example: "Essa é a primeira demanda"
            salario:
              type: "number"
              example: 5000.0
              format: "double"
            usuario:
              properties:
                idUsuario:
                  type: "number"
                  example: 1
        tecnologias:
          type: array
          example: [
          {
            tempoExperiencia: 10,
            tecnologias: {
              idTecnologia: 1
            }
          },
          {
            tempoExperiencia: 10,
            tecnologias: {
              idTecnologia: 2
            }
          }
          ]
        softskills:
          type: array
          example: [
          {
            notaSoftskill: 50,
            softskill: {
              idSoftSkills: 1
            }
          },
          {
            notaSoftskill: 50,
            softskill: {
              idSoftSkills: 2
            }
          }
          ]
    DemandaResponse:
      type: "array"
      example: [
        {
          idDemanda: 1,
          titulo: "Primeira demanda",
          descricao: "Essa é a primeira demanda",
          createdAt: "2021-06-04T14:51:18.553",
          salario: 0.0,
          usuario: {
              idUsuario: 3,
              nome: "Thiago Barbosa",
              email: "thiaguinho@gmail.com",
              senha: "senha123",
              descricao: "DBA",
              classificacao: 5.0,
              telefone: "962345673",
              localizacao: {
                  idLocalizacao: 1,
                  cidade: "São Caetano do Sul",
                  uf: "SP",
                  cep: "09572520"
              }
          }
        },
        {
          idDemanda: 2,
          titulo: "Primeira demanda",
          descricao: "Essa é a primeira demanda",
          createdAt: "2021-06-04T14:51:21.737",
          salario: 0.0,
          usuario: {
              idUsuario: 3,
              nome: "Thiago Barbosa",
              email: "thiaguinho@gmail.com",
              senha: "senha123",
              descricao: "DBA",
              classificacao: 5.0,
              telefone: "962345673",
              localizacao: {
                  idLocalizacao: 1,
                  cidade": "São Caetano do Sul",
                  uf: "SP",
                  cep: "09572520"
              }
          }
        }
        ]
    DemandaIdResponse:
      type: "object"
      properties:
        idDemanda: 
          type: number
          example: 3
        titulo: 
          type: string
          example: Segunda demanda
        descricao: 
          type: string
          example: Essa é a primeira demanda
        createdAt: 
          type: string
          example: 2021-06-05T16:27:57.36
        salario: 
          type: number
          example: 2000
        usuario: 
          type: object
          properties: 
            idUsuario: 
              type: number
              example: 3
            nome: 
              type: string
              example: Thiago Barbosa
            email: 
              type: string
              example: thiaguinho@gmail.com
            senha: 
              type: string
              example: senha123
            descricao: 
              type: string
              example: DBA
            classificacao: 
              type: number
              example: 5
            telefone: 
              type: string
              example: 962345673
            localizacao: 
              type: object
              properties: 
                idLocalizacao: 
                  type: number
                  example: 1
                cidade: 
                  type: string
                  example: São Caetano do Sul
                uf: 
                  type: string
                  example: SP
                cep: 
                  type: string
                  example: 09572520
    DemandaUsuarioResponse:
      type: "array"
      example: [
        {
          idDemanda: 1,
          titulo: "Primeira demanda",
          descricao: "Essa é a primeira demanda",
          createdAt: "2021-06-04T14:51:18.553",
          salario: 0.0,
          usuario: {
              idUsuario: 3,
              nome: "Thiago Barbosa",
              email: "thiaguinho@gmail.com",
              senha: "senha123",
              descricao: "DBA",
              classificacao: 5.0,
              telefone: "962345673",
              localizacao: {
                  idLocalizacao: 1,
                  cidade: "São Caetano do Sul",
                  uf: "SP",
                  cep: "09572520"
              }
          }
        },
        {
          idDemanda: 2,
          titulo: "Primeira demanda",
          descricao: "Essa é a primeira demanda",
          createdAt: "2021-06-04T14:51:21.737",
          salario: 0.0,
          usuario: {
              idUsuario: 3,
              nome: "Thiago Barbosa",
              email: "thiaguinho@gmail.com",
              senha: "senha123",
              descricao: "DBA",
              classificacao: 5.0,
              telefone: "962345673",
              localizacao: {
                  idLocalizacao: 1,
                  cidade": "São Caetano do Sul",
                  uf: "SP",
                  cep: "09572520"
              }
          }
        }
        ]
...
    DemandaFiltro:
      type: "object"
      properties:
        titulo: 
          type: string
          example: Primeira
        data: 
          type: string
          example: 2021-06-05T16:27:57.36
        UF: 
          type: string
          example: SP
        salarioMin: 
          type: number
          example: 2400
        salarioMax: 
          type: number
          example: 5000
        usuario: 
          type: string
          example: Thiago
        tecnologia: 
          type: string
          example: Java
        experiencia: 
          type: number
          example: 1
    DemandaFiltroResponse:
      type: "array"
      example: [
      {
        idDemanda: 8,
        titulo: Primeira demanda,
        descricao: Essa é a primeira demanda,
        createdAt: 2021-06-07T18:16:29.04,
        salario: 5000.0,
        usuario: {
            idUsuario: 3,
            nome: Thiago Barbosa,
            email: thiaguinho@gmail.com,
            senha: senha123,
            descricao: DBA,
            classificacao: 5.0,
            telefone: 962345673,
            localizacao: {
                idLocalizacao: 1,
                cidade: São Caetano do Sul,
                uf: SP,
                cep: 09572520
            }
        }
      },
      {
        idDemanda: 9,
        titulo: Primeira demanda,
        descricao: Essa é a primeira demanda,
        createdAt: 2021-06-07T18:20:26.303,
        salario: 5000.0,
        usuario: {
            idUsuario: 3,
            nome: Thiago Barbosa,
            email: thiaguinho@gmail.com,
            senha: senha123,
            descricao: DBA,
            classificacao: 5.0,
            telefone: 962345673,
            localizacao: {
                idLocalizacao: 1,
                cidade: São Caetano do Sul,
                uf: SP,
                cep: 09572520
            }
        }
      }
      ]
...
    ExportacaoDemanda:
      type: "object"
      properties: 
        titulo: 
          type: string
          example: Minha demanda
        descricao: 
          type: string
          example: Essa é a demanda que criei
        salario: 
          type: number
          example: 1000
        usuario:
          properties: 
            idUsuario:
              type: number
              example: 1
...
    OfertaIdResponse:
      type: "object"
      properties:
        idOferta: 
          type: number
          example: 3
        descricao: 
          type: string
          example: Essa é a primeira demanda
        createdAt: 
          type: string
          example: 2021-06-04T18:50:54.98
        usuario: 
          type: object
          properties: 
            idUsuario: 
              type: number
              example: 3
            nome: 
              type: string
              example: Thiago Barbosa
            email: 
              type: string
              example: thiaguinho@gmail.com
            senha: 
              type: string
              example: senha123
            descricao: 
              type: string
              example: DBA
            classificacao: 
              type: number
              example: 5
            telefone: 
              type: string
              example: 962345673
            localizacao: 
              type: object
              properties: 
                idLocalizacao: 
                  type: number
                  example: 1
                cidade: 
                  type: string
                  example: São Caetano do Sul
                uf: 
                  type: string
                  example: SP
                cep: 
                  type: string
                  example: 09572520
...
    Usuario:
      type: "object"
      properties: 
        nome: 
          type: string
          example: Mateus Gomes
        email: 
          type: string
          example: mateusgomes@gmail.com
        senha: 
          type: string
          example: senha123
        descricao: 
          type: string
          example: Backend Developer
        telefone: 
          type: string
          example: 962345673
        localizacao: 
          type: object
          properties: 
            idLocalizacao: 
              type: number
              example: 2
...
    UsuarioIdResponse:
      type: object
      properties:
        idUsuario: 
          type: number
          example: 3
        nome: 
          type: string
          example: Thiago Barbosa
        email: 
          type: string
          example: thiaguinho@gmail.com
        senha: 
          type: string
          example: senha123
        descricao: 
          type: string
          example: DBA
        classificacao: 
          type: number
          example: 5
        telefone: 
          type: string
          example: 962345673
        localizacao: 
          type: object
          properties: 
            idLocalizacao: 
              type: number
              example: 1
            cidade: 
              type: string
              example: São Caetano do Sul
            uf: 
              type: string
              example: SP
            cep: 
              type: string
              example: 09572520
...